refactored txn form components
